+++ include/s_stats.h
@@ -27,9 +27,9 @@
 #ifndef INCLUDED_s_stats_h
 #define INCLUDED_s_stats_h
 
-#define _1MEG     (1024.0)
-#define _1GIG     (1024.0*1024.0)
-#define _1TER     (1024.0*1024.0*1024.0)
+#define _1MEG	  (1024.0)
+#define _1GIG	  (1024.0*1024.0)
+#define _1TER	  (1024.0*1024.0*1024.0)
 #define _GMKs(x)  ( (x > _1TER) ? "Terabytes" : ((x > _1GIG) ? "Gigabytes" : \
 		  ((x > _1MEG) ? "Megabytes" : "Kilobytes")))
 #define _GMKv(x)  ( (x > _1TER) ? (float)(x/_1TER) : ((x > _1GIG) ? \
@@ -45,12 +45,12 @@ struct ServerStatistics
 	unsigned int is_cl;	/* number of client connections */
 	unsigned int is_sv;	/* number of server connections */
 	unsigned int is_ni;	/* connection but no idea who it was */
-	unsigned long long int is_cbs;	/* bytes sent to clients */
-	unsigned long long int is_cbr;	/* bytes received to clients */
-	unsigned long long int is_sbs;	/* bytes sent to servers */
-	unsigned long long int is_sbr;	/* bytes received to servers */
-	unsigned long long int is_cti;		/* time spent connected by clients */
-	unsigned long long int is_sti;		/* time spent connected by servers */
+	uint64_t is_cbs;	/* bytes sent to clients */
+	uint64_t is_cbr;	/* bytes received to clients */
+	uint64_t is_sbs;	/* bytes sent to servers */
+	uint64_t is_sbr;	/* bytes received to servers */
+	uint64_t is_cti;		/* time spent connected by clients */
+	uint64_t is_sti;		/* time spent connected by servers */
 	unsigned int is_ac;	/* connections accepted */
 	unsigned int is_ref;	/* accepts refused */
 	unsigned int is_unco;	/* unknown commands */
